tests/wpt/web-platform-tests/visual-viewport/viewport-segments.tentative.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>Viewport Segments: visualViewport.segments tentative</title>
<link rel="help" href="https://wicg.github.io/visual-viewport/"/>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
"use strict";

test(() => {
  assert_true('segments' in visualViewport, "`segments` must be a property of visualViewport.");
  assert_equals(visualViewport.segments, null, "For a viewport not segmented, no segments must be exposed.");
});
</script>
